+++ Modules/lootFrame.lua
@@ -101,7 +101,7 @@ function LootFrame:Update()
 		end
 	end
 	self.EntryManager:Update()
-	self.frame:SetHeight(numEntries * ENTRY_HEIGHT + 7)
+	self.frame.content:SetHeight(numEntries * ENTRY_HEIGHT + 7)
 end
 
 function LootFrame:OnRoll(entry, button)
@@ -146,7 +146,7 @@ do
 			entry.itemText:SetText(entry.item.link or "error")
 			entry.icon:SetNormalTexture(entry.item.texture or "Interface\\InventoryItems\\WoWUnknownItem01")
 			local typeText = addon:GetItemTypeText(item.link, item.subType, item.equipLoc, item.isTier, item.isRelic)
-			entry.itemLvl:SetText((entry.item.ilvl or 0).."  |cff7fffff"..typeText.."|r")
+			entry.itemLvl:SetText(addon:GetItemLevelText(entry.item.ilvl, entry.item.isTier).."  |cff7fffff"..typeText.."|r")
 			if addon.mldb.timeout then
 				entry.timeoutBar:SetMinMaxValues(0, addon.mldb.timeout or addon.db.profile.timeout)
 				entry.timeoutBar:Show()