+++ changelog.md
@@ -6,13 +6,16 @@
    + Note: RCLootCouncil cannot track if these items will be Warforged/Titanforged. Only the guaranteed minimum ilvl is used.
 * There's now one of each item type included in the test items.
 * The class icon can now be replaced with spec icon.
+* Filter buttons' text now change color to indicate a filter is active.
+* During tests all outputs are now preceeded with "(test)".
+* Any chat output during solo tests are now replaced with chat prints.
 
 
 * **Localization**
-+ Armor Token is no longer displayed as "Junk" for non-English client.
++ Armor Token is no longer displayed as "Junk" for non-English clients.
 + Relics are now shown as "'type' Artifact Relic" instead of 'type'.
    + Relic type is now also shown for non-English client.
-+ Trinkets are now shown as "Trinket" instead of "Trinket, Miscellaneous" for non-English client
++ Trinkets are now shown as "Trinket" instead of "Trinket, Miscellaneous" for non-English client.
 
 
 * **Announcements**
@@ -26,6 +29,8 @@
    + E.g. the council can now see a candidates gear and ilvl before a candidates responds.
 * RCLootCouncil now sends the gear a candidate had equipped during the most recent encounter instead of the gear equipped when rolling.
    + This way candidates can't change their gear to appear to have lower ilvl.
+* If enabled, relic responses can now be filtered separately.
+* You can now filter responses from candidates that can't use a given item.
 
 
 * **Rolls**
@@ -55,6 +60,7 @@
    + ```:UpdatePlayersData()``` uses the two new functions ```:UpdatePlayersRelics()``` and ```:UpdatePlayerGears()``` to cache the player's gear/info.
    + ```:SendResponse(target, session, link, ilvl, response, equipLoc, note, subType, isTier, isRelic, sendAvgIlvl)``` sends response to the target.
    + SpecID is now included in the candidate data.
+   + New message on ```:BuildMLdb()```.
 
 * **Changed**
    + ~~```:CreateResponse()```~~ is consolidated into ```:SendResponse(...)``` which now creates and sends responses.